refactor: reorganize database service classes into a dedicated package
