+++ vulcanus/src/main/java/com/merkury/vulcanus/db/services/PopulateSpotsService.java
@@ -1,0 +1,9 @@
+package com.merkury.vulcanus.db.services;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class PopulateSpotsService {
+}
